src/etl_pipeline.py
import torch
from torchvision import datasets, transforms
from nvidia.dali import pipeline, ops, types
import os

class DALIPipeline:
    """Pipeline de ETL acelerado por GPU usando NVIDIA DALI"""
    
    def __init__(self, batch_size=64, device_id=0, num_threads=4):
        self.batch_size = batch_size
        self.device_id = device_id
        self.num_threads = num_threads
        
    def create_data_pipeline(self, data_path):
        """Cria pipeline de pré-processamento com DALI"""
        pipe = pipeline.Pipeline(
            batch_size=self.batch_size,
            num_threads=self.num_threads,
            device_id=self.device_id
        )
        
        with pipe:
            # Para dados reais, usar FileReader
            jpegs, labels = ops.external_source(
                source=self.get_data_generator(data_path),
                num_outputs=2,
                device="cpu"
            )
            
            images = ops.ImageDecoder(
                device="mixed",
                output_type=types.RGB
            )(jpegs)
            
            images = ops.Resize(
                device="gpu",
                resize_x=32,
                resize_y=32
            )(images)
            
            output = ops.CropMirrorNormalize(
                device="gpu",
                output_dtype=types.FLOAT,
                output_layout=types.NCHW,
                mean=[0.485 * 255, 0.456 * 255, 0.406 * 255],
                std=[0.229 * 255, 0.224 * 255, 0.225 * 255]
            )(images)
            
            pipe.set_outputs(output, labels)
        
        return pipe
    
    def get_data_generator(self, data_path):
        """Generator para simular stream de dados"""
        # Implementar generator real para seus dados
        pass

def download_dataset():
    """Baixa e prepara dataset CIFAR-10"""
    transform = transforms.Compose([
        transforms.ToTensor(),
        transforms.Normalize((0.5, 0.5, 0.5), (0.5, 0.5, 0.5))
    ])
    
    train_dataset = datasets.CIFAR10(
        root='./data',
        train=True,
        download=True,
        transform=transform
    )
    
    print(f"Dataset baixado: {len(train_dataset)} imagens")
    return train_dataset

if __name__ == "__main__":
    # Exemplo de uso
    dataset = download_dataset()
    print("ETL pipeline configurado com sucesso!")